New stable release build

+++ shared/http/http_incoming_request.cpp
@@ -1,7 +1,6 @@
-#include "http_incoming_request.h"
 //
 // This file is part of the WoW++ project.
-// 
+//
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
 // the Free Software Foundation; either version 2 of the License, or
@@ -11,20 +10,20 @@
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 // GNU General Public License for more details.
-// 
+//
 // You should have received a copy of the GNU General Public License
-// along with this program; if not, write to the Free Software 
+// along with this program; if not, write to the Free Software
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 //
 // World of Warcraft, and all World of Warcraft or Warcraft art, images,
 // and lore are copyrighted by Blizzard Entertainment, Inc.
-// 
+//
 
+#include "pch.h"
 #include "http_incoming_request.h"
 #include "common/constants.h"
 #include "base64/base64.h"
 #include "log/default_log_levels.h"
-#include <cassert>
 
 namespace wowpp
 {
@@ -52,7 +51,7 @@ namespace wowpp
 				return m_host;
 			}
 
-			const String & IncomingRequest::getPostData() const
+			const String &IncomingRequest::getPostData() const
 			{
 				return m_postData;
 			}
@@ -236,7 +235,7 @@ namespace wowpp
 				if (estimatedContentLength > 0)
 				{
 					if (end <= pos ||
-						(end - pos) < estimatedContentLength)
+					        size_t(end - pos) < estimatedContentLength)
 					{
 						return receive_state::Incomplete;
 					}
